Changed version_h.pl to only rewrite version.h when the revision number changes.
git-svn-id: http://svn.pugscode.org/pugs@4724 c213334d-75ef-0310-aa23-eaa082d1ae64

+++ util/version_h.pl
@@ -7,9 +7,17 @@ my $version_h = shift || Cwd::cwd() . "/src/Pugs/pugs_version.h";
 my $base = shift || Cwd::cwd();
 my $svn_entries = "$base/.svn/entries";
 
+my $old_revision = 0;
+open IN, "< $version_h" and do {
+  while (<IN>) {
+    /#define PUGS_SVN_REVISION (\d+)/ or next;
+    $old_revision = $1;
+    last;
+  }
+  close IN;
+};
+
 my $revision = 0;
-open OUT, "> $version_h" or die $!;
-print OUT "#undef PUGS_SVN_REVISION\n";
 
 # SVK tries to ask the user questions when it has a STDIN and there is
 # no repository.  Since we don't need a STDIN anyway, get rid of it.
@@ -30,12 +38,19 @@ if (-r $svn_entries) {
     ($revision) = $line =~ / (\d+)$/;
 }
 $revision ||= 0;
-print OUT "#define PUGS_SVN_REVISION $revision\n";
-close OUT;
 
-if ($revision != 0) {
+if ($revision != $old_revision) {
+  open OUT, "> $version_h" or die $!;
+  print OUT "#undef PUGS_SVN_REVISION\n";
+  print OUT "#define PUGS_SVN_REVISION $revision\n";
+  close OUT;
+
+  if ($revision != 0) {
     # rebuild Help.hs to show new revision number
     unlink "$base/src/Pugs/Help.hi";
     unlink "$base/src/Pugs/Help.o";
     exit;
+  }
+} else {
+  print "Not writing $version_h because $old_revision == $revision\n";
 }
